print xml parse errors to stdout in silvia_irma_xmlreader
Ideally, this should be done in a different way (either through a callback, an exception or something like strerror), but since libxml itself can also print errors to stderr, I kept the structure of the code intact.

Alternatively, we may want to think about using a schema validator instead of hand-picking elements.